Hand TabPFN fp32 under autocast; smoke the bf16 path on one gpu
v5 died at step 0 with "Got unsupported ScalarType BFloat16":
AddFingerprintFeaturesStep hashes its input via X.detach().cpu().numpy() and
numpy has no bf16, so the adapter's autocast-emitted bf16 output cannot enter
TabPFN's preprocessing. Casting back to fp32 at the adapter boundary costs
nothing -- autocast casts per op, so the transformer's own projections are
still bf16 and flash attention is still eligible.

My earlier claim that the differentiable pipeline is only RemoveConstantFeatures
+ DifferentiableZNorm was wrong; the fingerprint step runs too.

Submitting a 1-gpu, 2-step, 8-task smoke first: a dtype error at step 0 costs a
4-gpu slot, and the last one took 25 minutes of queue to get.

+++ expts/repaper/adapter/train_adapter.py
@@ -298,7 +298,14 @@ def loss_and_pred(ests, adapter, task_type, emb, y, n_ctx, device, autocast_dtyp
         else contextlib.nullcontext()
     )
     with ctx:
-        x = adapter(emb.to(device).float())
+        # .float() back immediately: autocast makes the adapter's own Linear
+        # emit bf16, and TabPFN's AddFingerprintFeaturesStep hashes its input
+        # via X.detach().cpu().numpy(), which raises "unsupported ScalarType
+        # BFloat16" because numpy has no bf16. Handing TabPFN fp32 costs
+        # nothing -- autocast casts per *op*, so the projections inside the
+        # transformer are still bf16 and flash attention is still eligible;
+        # only this one 512x512 matmul stays fp32.
+        x = adapter(emb.to(device).float()).float()
         yy = y.to(device)
         est = ests[task_type]
         # A task can vary over its whole pool and still draw a context that
